+++ Classes/StrategiaConteggio.cs
@@ -18,7 +18,7 @@ namespace Classes
             Punteggio = punteggio;
         }
 
-        public override Giocatore.Puntata Strategy(Giocatore giocatore, Mazziere mazziere, double conteggio)
+        public override Giocatore.Puntata Strategy(Giocatore giocatore, Mazziere mazziere, decimal conteggio)
         {
             if (giocatore.Punteggio < Punteggio)
                 return Giocatore.Puntata.Chiama;
@@ -26,7 +26,7 @@ namespace Classes
                 return Giocatore.Puntata.Stai;
         }
 
-        public override double Puntata(Giocatore giocatore, double puntataMinima, double puntataBase, double Conteggio)
+        public override decimal Puntata(Giocatore giocatore, decimal puntataMinima, decimal puntataBase, decimal Conteggio)
         {
             if (Conteggio <= -2) //Lasciare il tavolo
                 return puntataMinima;
@@ -50,7 +50,7 @@ namespace Classes
             return Conteggio;
         }
 
-        public override bool Assicurazione(Giocatore giocatore, double conteggio)
+        public override bool Assicurazione(Giocatore giocatore, decimal conteggio)
         {
             return false;
         }
